refactored cli modules
