+++ gpt_engineer/api.py
@@ -2,7 +2,7 @@ import os
 from pathlib import Path
 from agent_protocol import Agent, Step, Task, models
 from gpt_engineer.core.db import DB
-from gpt_engineer.main import main
+from gpt_engineer.cli.main import main
 from openai.error import AuthenticationError
 import tempfile
