+++ gpt_engineer/core/chat_to_files.py
@@ -31,7 +31,7 @@ import re
 from typing import List, Tuple
 
 from gpt_engineer.core.db import DB, DBs
-from gpt_engineer.file_selector import FILE_LIST_NAME
+from gpt_engineer.cli.file_selector import FILE_LIST_NAME
 
 
 def parse_chat(chat) -> List[Tuple[str, str]]:
